# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Liberfly",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get a JWT via credentials",
                "description": "Returns a JWT",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "admin@admin.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "password"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Wrong credentials"
                    },
                    "403": {
                        "description": "The user is forbidden"
                    },
                    "404": {
                        "description": "User not found or wrong credentials"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated User.",
                "description": "Get the authenticated User.",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "The user is not authenticated."
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "Logoutt User",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "The user is not authenticated."
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh a Token",
                "description": "Refresh a Token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "The user is not authenticated."
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Return Products List.",
                "description": "Return Products List.",
                "operationId": "eb3eee756c2994d4ff825b022be31b89",
                "responses": {
                    "200": {
                        "description": "Products List"
                    },
                    "404": {
                        "description": "Route Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/store": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a product",
                "description": "Create a product",
                "operationId": "b331f25302ee0779e4ff1e46b2d10224",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nome do Produto"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Televisores"
                                    },
                                    "price": {
                                        "type": "decimal",
                                        "example": "50.49"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/show/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Returno Products List.",
                "description": "Returno Products List.",
                "operationId": "17f7dbce03df6f0887a8cad872651cd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products List"
                    },
                    "404": {
                        "description": "user not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/destroy/{id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "description": "Delete a product",
                "operationId": "c839884bb38bb4867a7c23f5f5f3363a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}